GetLakeMap

Function to retrieve the spatial coordinates of Lake from the AK_ShallowLakes:tblMonuments table and display them in a Leaflet map.

Usage

GetLakeMap(Lake, ZoomLevel)

Arguments

Lake

Lake for which spatial coordinates should be returned

ZoomLevel

Initial zoom level for the lake map. Around 12-15 is usually appropriate. The user may change the zoom level when after the map is generated.

Value

A Leaflet map for Lake.

Examples

# Get a Leaflet map showing Lake YUCH-004 at zoom level 13
GetLakeMap('YUCH-004',13)
#> Error: ! ODBC failed with error 00000 from [unixODBC][Driver Manager].
#> ✖ Can't open lib 'Sql Server' : file not found
#> ℹ From nanodbc/nanodbc.cpp:1150.
#> Error: unable to find an inherited method for function ‘dbGetQuery’ for signature ‘conn = "logical", statement = "character"’