GetWaterProfilesDataSummary

The GetWaterProfilesDataSummary function returns a summary of water profile parameters over sample replicates (usually A,B,C,etc.) and grouped by Lake and sampling date. Statistics include mean, min, max, SD and n) for each lake and sampling date. The statistics returned by this function are calculated by the AK_ShallowLakes database in the view (query) [Summary_WaterProfileStatisticsByLake]. Results are returned as a data frame.

Usage

GetWaterProfilesDataSummary()

Value

A data frame of water profile statistics for temperature, pH, and dissolved oxygen.

Examples

# Example of use:
data = GetWaterProfilesDataSummary
head(data)
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> 1 function ()                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> 2 {                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> 3     Connection <- tryCatch({                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#> 4         Connection = GetDatabaseConnection()                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#> 5         Sql = " SELECT [PONDNAME]\\n,[SAMPLEDATE]\\n,Month(SampleDate) as [Month]\\n,Case When Month(SampleDate) < 7 Then 'Spring' Else 'Fall' End as Season\\n,[SAMPLEDEPTH]\\n,[Temperature_Mean]\\n,[Temperature_Min]\\n,[Temperature_Max]\\n,[Temperature_SD]\\n,[Temperature_n]\\n,[pH_Mean]\\n,[pH_Min]\\n,[pH_Max]\\n,[pH_SD]\\n,[pH_n]\\n,[SpCond_Mean]\\n,[SpCond_Min]\\n,[SpCond_Max]\\n,[SpCond_SD]\\n,[SpCond_n]\\n,[DO_Mean]\\n,[DO_Min]\\n,[DO_Max]\\n,[DO_SD]\\n,[DO_n]\\n,[DO_Pct_Mean]\\n,[DO_Pct_Min]\\n,[DO_Pct_Max]\\n,[DO_Pct_SD]\\n,[DO_Pct_n]\\nFROM [AK_ShallowLakes].[dbo].[Summary_WaterProfileStatisticsByLake]\\nORDER BY PONDNAME,SAMPLEDATE,SAMPLEDEPTH"
#> 6         DF = odbc::dbGetQuery(Connection, Sql)